fix css navbar confirmTicket searchBar mobile et desktop
