+++ frontend/src/components/Articles/Article.scss
@@ -33,7 +33,7 @@ body {
 .dateArticle {
   position: absolute;
   padding-top: 112px;
-  padding-left: 80px;
+  padding-left: 46px;
   font-style: italic;
   font-size: 10px;
   letter-spacing: 0.1em;
