+++ frontend/src/components/Compteurcard/Card.scss
@@ -77,6 +77,7 @@ $breakpoints: (
 
 .infos-card {
   display: flex;
+  width: 60%;
   flex-direction: column;
   align-items: center;
   border-radius: 10px;
