+++ frontend/src/pages/Home_FlyNews.scss
@@ -8,6 +8,15 @@ $color-background: #1e1e1e;
 $color-shadows-dark: #101010;
 $color-light: #d7d7d7;
 $color-dark: #221e40;
+$rad: 0.7rem;
+$dur: 0.3s;
+$color-dark: #2f2f2f;
+$color-light: #fff;
+$color-brand: #57bd84;
+$font-fam: "Lato", sans-serif;
+$height: 5rem;
+$btn-width: 50px;
+$bez: cubic-bezier(0, 0, 0.43, 1.49);
 
 $breakpoints: (
   mobile: 600px,
@@ -42,16 +51,30 @@ $breakpoints: (
   display: flex;
   flex-direction: row;
   margin: 1% 0% -2% 1%;
+  width: 12rem;
+  background-color: $color-Bck-btn;
+  border-radius: 0.7rem;
+  margin-left: 30px;
 }
 .search-label {
-  margin: 2% 0% 2% 2%;
-  font-family: "Inria Sans", sans-serif;
-  color: #4d51aa;
-  font-size: 20px;
-  font-weight: bold;
+  color: $color-Bck-div;
+  font-size: 26px;
+  padding: 2px;
 }
 .search-input {
-  margin: 2%;
-  background-color: #d7d7d7;
-  color: $color-dark;
+  outline: 0; // <-- shold probably remove this for better accessibility, adding for demo aesthetics for now.
+  width: 100%;
+  background: $color-light;
+  padding: 0 1.6rem;
+  border-radius: $rad;
+  appearance: none; //for iOS input[type="search"] roundedness issue. border-radius alone doesn't work
+  transition: all $dur $bez;
+  transition-property: width, border-radius;
+  z-index: 1;
+  position: relative;
+}
+
+.search-input:not(:placeholder-shown) {
+  color: $color-Bck-btn;
+  width: calc(100% + $btn-width);
 }
